replace useNetworkd with nixos-generate-config does
There is the caveat that interface names could potentially differ
between an installer and the installed system.

Overall I would like to burn scripted networking with fire, but this is not the
place and time to do it.

Ideally we would also use mac addresses to make sure sure interface
names are stable, but don't have this information in facter report for
privacy reasons.